Restructured Repository#cached_commits
Use lambdas instead of recursion, so the already loaded commits are
known and potential cycles in the history can be found.

Additionally, cached commits are now sorted:
Complex merged histories cause commits to be in a more or less random
order. But order is important when we determine the head / tail of the
history.

Closes #17.